requirements.txt
greenlet==1.1.2
PyMySQL==1.0.2
python-dotenv==0.19.2
SQLAlchemy==1.4.31
